requirement.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.8.1
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.9.0
Django==5.0.7
django-browser-reload==1.13.0
django-tailwind==3.8.0
djangorestframework==3.15.2
dnspython==2.5.0
email-validator==2.1.0.post1
emails==0.6
fastapi==0.109.0
fastapi-mail==1.4.1
firebase-admin==6.5.0
flake8==7.0.0
Flask==3.0.1
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.4
mccabe==0.7.0
motor==3.3.2
msgpack==1.0.8
orjson==3.9.13
premailer==3.10.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.0
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
pyflakes==3.2.0
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
secure-smtplib==0.1.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.5.1
starlette==0.35.1
typing_extensions==4.9.0
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.2.0
uv==0.2.30
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
